+++ code_1.1/code_1.X/dac_10bit.h
@@ -5,7 +5,7 @@
 #include <stdint.h> 
 
 /* Function prototypes (wrapper functions) */
-float defuzzify_10bit (float x);
+uint16_t defuzzify_10bit (float x);
     
 #endif
